fix(backup): ordina il restore per dipendenze
Estrae l'esecutore di ripristino dal route handler e inserisce gli item di prestazione prima delle osservazioni collegate.
